Altered the Legal Notice.
Due to a bug in current Publican, building a book with the XREF element
in the Legal Notice can cause a memory leak. Consequent to this, the
book cannot be built, and Publican has to be killed from the command
line. As a workaround, I have removed the XREF, so that the book can be
built as expected.